requirements.txt
anyio==3.6.1
attrs==21.4.0
click==8.1.3
h11==0.13.0
idna==3.3
iniconfig==1.1.1
packaging==21.3
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pydantic==1.9.1
pyparsing==3.0.9
pytest==7.1.2
python-dotenv==0.20.0
sniffio==1.2.0
sqlparse==0.4.2
starlette==0.20.3
tabulate==0.8.10
tomli==2.0.1
typing_extensions==4.2.0
uvicorn==0.18.1
yoyo-migrations==7.3.2
